requirements.txt
# FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis for caching and queues
redis==5.0.1
celery==5.3.6

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography==42.0.0

# Image Processing
Pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.3
scikit-image==0.22.0

# Metadata Extraction
exifread==3.0.0
piexif==1.1.3

# C2PA Verification
c2pa-python==0.4.0

# ML and AI Detection
torch==2.1.2
torchvision==0.16.2
transformers==4.37.0
timm==0.9.12
huggingface-hub==0.20.3

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Utilities
python-magic==0.4.27
filetype==1.2.0
python-dateutil==2.8.2
pytz==2023.3

# Logging and Monitoring
structlog==24.1.0
sentry-sdk==1.39.2
prometheus-client==0.19.0

# Testing (also in requirements-dev.txt)
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Rate Limiting
slowapi==0.1.9

# Cloud Storage (AWS/GCP)
boto3==1.34.25
google-cloud-storage==2.14.0

# Background Tasks
apscheduler==3.10.4
